refactor/perf: flatten Object into Value enum
This way we can add a Rc<> only for the variants we need, e.g. String is
already reference counted internally, so we don't need another Rc around
it. NativeFun is just a function pointer, so we don't need a reference
count there either.

This also means we don't need to copy the chunk every time we push a
CallFrame, since now the CallFrame contains Rc<Function>.